extract queued events tests
