+++ tests/Events/EventsDispatcherTest.php
@@ -4,11 +4,7 @@ namespace Illuminate\Tests\Events;
 
 use Exception;
 use Illuminate\Container\Container;
-use Illuminate\Contracts\Queue\Queue;
-use Illuminate\Contracts\Queue\ShouldQueue;
-use Illuminate\Events\CallQueuedListener;
 use Illuminate\Events\Dispatcher;
-use Illuminate\Support\Testing\Fakes\QueueFake;
 use Mockery as m;
 use PHPUnit\Framework\TestCase;
 
@@ -318,39 +314,6 @@ class EventsDispatcherTest extends TestCase
         $d->dispatch('foo.bar', ['first', 'second']);
     }
 
-    public function testQueuedEventHandlersAreQueued()
-    {
-        $d = new Dispatcher;
-        $queue = m::mock(Queue::class);
-
-        $queue->shouldReceive('connection')->once()->with(null)->andReturnSelf();
-
-        $queue->shouldReceive('pushOn')->once()->with(null, m::type(CallQueuedListener::class));
-
-        $d->setQueueResolver(function () use ($queue) {
-            return $queue;
-        });
-
-        $d->listen('some.event', TestDispatcherQueuedHandler::class.'@someMethod');
-        $d->dispatch('some.event', ['foo', 'bar']);
-    }
-
-    public function testCustomizedQueuedEventHandlersAreQueued()
-    {
-        $d = new Dispatcher;
-
-        $fakeQueue = new QueueFake(new Container());
-
-        $d->setQueueResolver(function () use ($fakeQueue) {
-            return $fakeQueue;
-        });
-
-        $d->listen('some.event', TestDispatcherConnectionQueuedHandler::class.'@handle');
-        $d->dispatch('some.event', ['foo', 'bar']);
-
-        $fakeQueue->assertPushedOn('my_queue', CallQueuedListener::class);
-    }
-
     public function testClassesWork()
     {
         unset($_SERVER['__event.test']);
@@ -430,28 +393,6 @@ class EventsDispatcherTest extends TestCase
 
         $d->subscribe($subs);
         $this->assertTrue(true);
-    }
-}
-
-class TestDispatcherQueuedHandler implements ShouldQueue
-{
-    public function handle()
-    {
-        //
-    }
-}
-
-class TestDispatcherConnectionQueuedHandler implements ShouldQueue
-{
-    public $connection = 'redis';
-
-    public $delay = 10;
-
-    public $queue = 'my_queue';
-
-    public function handle()
-    {
-        //
     }
 }
