+++ tests/Events/QueuedEventsTest.php
@@ -1,0 +1,75 @@
+<?php
+
+namespace Illuminate\Tests\Events;
+
+use Illuminate\Container\Container;
+use Illuminate\Contracts\Queue\Queue;
+use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Events\CallQueuedListener;
+use Illuminate\Events\Dispatcher;
+use Illuminate\Support\Testing\Fakes\QueueFake;
+use Mockery as m;
+use PHPUnit\Framework\TestCase;
+
+class QueuedEventsTest extends TestCase
+{
+    protected function tearDown(): void
+    {
+        m::close();
+    }
+
+    public function testQueuedEventHandlersAreQueued()
+    {
+        $d = new Dispatcher;
+        $queue = m::mock(Queue::class);
+
+        $queue->shouldReceive('connection')->once()->with(null)->andReturnSelf();
+
+        $queue->shouldReceive('pushOn')->once()->with(null, m::type(CallQueuedListener::class));
+
+        $d->setQueueResolver(function () use ($queue) {
+            return $queue;
+        });
+
+        $d->listen('some.event', TestDispatcherQueuedHandler::class.'@someMethod');
+        $d->dispatch('some.event', ['foo', 'bar']);
+    }
+
+    public function testCustomizedQueuedEventHandlersAreQueued()
+    {
+        $d = new Dispatcher;
+
+        $fakeQueue = new QueueFake(new Container());
+
+        $d->setQueueResolver(function () use ($fakeQueue) {
+            return $fakeQueue;
+        });
+
+        $d->listen('some.event', TestDispatcherConnectionQueuedHandler::class.'@handle');
+        $d->dispatch('some.event', ['foo', 'bar']);
+
+        $fakeQueue->assertPushedOn('my_queue', CallQueuedListener::class);
+    }
+}
+
+class TestDispatcherQueuedHandler implements ShouldQueue
+{
+    public function handle()
+    {
+        //
+    }
+}
+
+class TestDispatcherConnectionQueuedHandler implements ShouldQueue
+{
+    public $connection = 'redis';
+
+    public $delay = 10;
+
+    public $queue = 'my_queue';
+
+    public function handle()
+    {
+        //
+    }
+}
